feat(ui): land upstream ui_expressions_logicfunctions
Wire src/ui/ui_expressions_logicfunctions.cpp (1377 lines) into both
backends — the per-operator implementations for the menu-script
expression evaluator (compare, add, mul, bitwise, string concat, etc.).

Source fixes for -Werror clean build:
  - KISAKHACK (uintptr_t) bridge on all `(T *)<int_expr>` casts where
    Operand::internals.intVal is being dereferenced as a pointer
    (the 32-bit upstream type-puns a string pointer through an int)
  - KISAKHACK (int)(uintptr_t)<ptr> on the resultStr write-back into
    Operand::internals.intVal so the 32-bit value preserves the low
    word of the actual 64-bit pointer
  - KISAKHACK on `*(_BYTE *)<int_expr>` dereferences in bitwise/logical
    operand-of-operand chains

Satellite stubs in src/posix/posix_backbone_stubs.cpp:
  - GetSourceInt(Operand *) (returns operandInternalDataUnion per
    ui_shared.h, not a plain int)

Dup removal:
  - Comment out compare_hudelems and compare_impact_files now provided
    by ui_expressions_logicfunctions

+++ src/ui/ui_expressions_logicfunctions.cpp
@@ -9,7 +9,7 @@ int __cdecl compare_hudelems(const void *pe0, const void *pe1)
 {
     float delta; // [esp+0h] [ebp-Ch]
 
-    delta = *(float *)(*(unsigned int *)pe0 + 128) - *(float *)(*(unsigned int *)pe1 + 128);
+    delta = *(float *)(uintptr_t)(*(unsigned int *)pe0 + 128) - *(float *)(uintptr_t)(*(unsigned int *)pe1 + 128);
     if (delta >= 0.0)
         return delta > 0.0;
     else
@@ -597,10 +597,10 @@ void __cdecl add_StringWithString(Operand *leftSide, Operand *rightSide, Operand
             "(rightSide->dataType == VAL_STRING)",
             rightSide->dataType);
     result->dataType = VAL_STRING;
-    I_strncpyz(leftSideStr, (char *)leftSide->internals.intVal, 256);
-    I_strncpyz(rightSideStr, (char *)rightSide->internals.intVal, 256);
+    I_strncpyz(leftSideStr, (char *)(uintptr_t)leftSide->internals.intVal, 256);
+    I_strncpyz(rightSideStr, (char *)(uintptr_t)rightSide->internals.intVal, 256);
     Com_sprintf(resultStr, 0x100u, "%s%s", leftSideStr, rightSideStr);
-    result->internals.intVal = (int)resultStr;
+    result->internals.intVal = (int)(uintptr_t)resultStr;
 }
 
 char resultStr_0[256];
@@ -625,9 +625,9 @@ void __cdecl add_StringWithInt(Operand *leftSide, Operand *rightSide, Operand *r
             "(rightSide->dataType == VAL_INT)",
             rightSide->dataType);
     result->dataType = VAL_STRING;
-    I_strncpyz(leftSideStr, (char *)leftSide->internals.intVal, 256);
+    I_strncpyz(leftSideStr, (char *)(uintptr_t)leftSide->internals.intVal, 256);
     Com_sprintf(resultStr_0, 0x100u, "%s%i", leftSideStr, rightSide->internals.intVal);
-    result->internals.intVal = (int)resultStr_0;
+    result->internals.intVal = (int)(uintptr_t)resultStr_0;
 }
 
 char resultStr_1[256];
@@ -652,9 +652,9 @@ void __cdecl add_IntWithString(Operand *leftSide, Operand *rightSide, Operand *r
             "(rightSide->dataType == VAL_STRING)",
             leftSide->dataType);
     result->dataType = VAL_STRING;
-    I_strncpyz(rightSideStr, (char *)rightSide->internals.intVal, 256);
+    I_strncpyz(rightSideStr, (char *)(uintptr_t)rightSide->internals.intVal, 256);
     Com_sprintf(resultStr_1, 0x100u, "%i%s", leftSide->internals.intVal, rightSideStr);
-    result->internals.intVal = (int)resultStr_1;
+    result->internals.intVal = (int)(uintptr_t)resultStr_1;
 }
 
 char resultStr_2[256];
@@ -679,9 +679,9 @@ void __cdecl add_FloatWithString(Operand *leftSide, Operand *rightSide, Operand 
             "(rightSide->dataType == VAL_STRING)",
             leftSide->dataType);
     result->dataType = VAL_STRING;
-    I_strncpyz(rightSideStr, (char *)rightSide->internals.intVal, 256);
+    I_strncpyz(rightSideStr, (char *)(uintptr_t)rightSide->internals.intVal, 256);
     Com_sprintf(resultStr_2, 0x100u, "%f%s", leftSide->internals.floatVal, rightSideStr);
-    result->internals.intVal = (int)resultStr_2;
+    result->internals.intVal = (int)(uintptr_t)resultStr_2;
 }
 
 char resultStr_3[256];
@@ -706,9 +706,9 @@ void __cdecl add_StringWithFloat(Operand *leftSide, Operand *rightSide, Operand 
             "(rightSide->dataType == VAL_FLOAT)",
             rightSide->dataType);
     result->dataType = VAL_STRING;
-    I_strncpyz(leftSideStr, (char *)leftSide->internals.intVal, 256);
+    I_strncpyz(leftSideStr, (char *)(uintptr_t)leftSide->internals.intVal, 256);
     Com_sprintf(resultStr_3, 0x100u, "%s%f", leftSideStr, rightSide->internals.floatVal);
-    result->internals.intVal = (int)resultStr_3;
+    result->internals.intVal = (int)(uintptr_t)resultStr_3;
 }
 
 void __cdecl multiply_IntByInt(Operand *leftSide, Operand *rightSide, Operand *result)
@@ -1064,7 +1064,7 @@ void __cdecl and_StringWithInt(Operand *leftSide, Operand *rightSide, Operand *r
             "(rightSide->dataType == VAL_INT)",
             rightSide->dataType);
     result->dataType = VAL_INT;
-    v3 = *(_BYTE *)leftSide->internals.intVal && rightSide->internals.intVal;
+    v3 = *(_BYTE *)(uintptr_t)leftSide->internals.intVal && rightSide->internals.intVal;
     result->internals.intVal = v3;
 }
 
@@ -1094,7 +1094,7 @@ void __cdecl and_StringWithFloat(Operand *leftSide, Operand *rightSide, Operand 
             "(rightSide->dataType == VAL_FLOAT)",
             rightSide->dataType);
     result->dataType = VAL_INT;
-    v3 = *(_BYTE *)leftSide->internals.intVal && rightSide->internals.floatVal != 0.0;
+    v3 = *(_BYTE *)(uintptr_t)leftSide->internals.intVal && rightSide->internals.floatVal != 0.0;
     result->internals.intVal = v3;
 }
 
@@ -1204,7 +1204,7 @@ void __cdecl or_StringWithInt(Operand *leftSide, Operand *rightSide, Operand *re
             "(rightSide->dataType == VAL_INT)",
             rightSide->dataType);
     result->dataType = VAL_INT;
-    v3 = *(_BYTE *)leftSide->internals.intVal || rightSide->internals.intVal;
+    v3 = *(_BYTE *)(uintptr_t)leftSide->internals.intVal || rightSide->internals.intVal;
     result->internals.intVal = v3;
 }
 
@@ -1234,7 +1234,7 @@ void __cdecl or_StringWithFloat(Operand *leftSide, Operand *rightSide, Operand *
             "(rightSide->dataType == VAL_FLOAT)",
             rightSide->dataType);
     result->dataType = VAL_INT;
-    v3 = *(_BYTE *)leftSide->internals.intVal || rightSide->internals.floatVal != 0.0;
+    v3 = *(_BYTE *)(uintptr_t)leftSide->internals.intVal || rightSide->internals.floatVal != 0.0;
     result->internals.intVal = v3;
 }
